+++ src/Resources/contao/languages/en/tl_content.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * Contao Bootstrap templates.
+ *
+ * @package    contao-bootstrap
+ * @subpackage Templates
+ * @author     David Molineus <david.molineus@netzmacht.de>
+ * @copyright  2014-2019 netzmacht David Molineus. All rights reserved.
+ * @license    https://github.com/contao-bootstrap/templates/blob/master/LICENSE LGPL 3.0-or-later
+ * @filesource
+ */
+
+declare(strict_types=1);
+
+
+$GLOBALS['TL_LANG']['tl_content']['playerAspect'][0] = 'Aspect ratio';
+$GLOBALS['TL_LANG']['tl_content']['playerAspect'][1] = 'Here you can select the aspect ratio of the video to make it responsive.';