+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog
 
  - Provide newsletter template `nl_bs`
  - Button classes are insert at a tag for hyperlink elements (See #22)
+ - Add responsive-embed classes for youtube and vimeo player having aspect ratio ([#21](https://github.com/contao-bootstrap/templates/issues/21))
+ - Backport playerAspect for video players so that responsive-embed is usable in Contao 4.4 ([#21](https://github.com/contao-bootstrap/templates/issues/21))
 
 2.0.0 (2018-08-01)
 ------------------